Added framework for writing functional tests
The new conftest now includes a ``srcds`` fixtures which allows
Source server instances to be ran.

The envrionment variable PYTHON_VALVE_SRCDS must be set to a
directory containing an instance/ directory which contains a
bunch of SRCDS installations.

By default the tests using ``srcds`` fixture won't be run unless
the --srcds command line option is set. This option should be
a comma-separated list of SRCDS instances (the names of the dirs
in $PYTHON_VALVE_SRCDS/instances/). The tests will be ran against
all instances specified. If --srcds=all then all tests will run
against all discovered instances.

Biggest flaw with this at the moment is that SRCDSRunner has no
way to tell when the server it 'up', because *issues* arose when
trying to scan stdout for landmarks. :(