A2S_INFO's 'server_type' field now returns new ServerType instances
The ServerType class performs the same function as Platform except
for server types. Current server types are "Dedicated," "Non-Dedicated"
and "SourceTV." Non-dedicated is represented by ord("l") which leads me
to believe it infact means 'listen server' Will have to investigate to
determine if this is the case. If it is the non-dedicated type will
be renamed.