Initial stab at RCON authentication
Interestingly, RCON servers (at least the TF2 I tested against) will send
an initial 'empty' message. As in, a message of size 14 but everything other
than the first byte of the size field is null.

After that blank message it then sends the actual AUTH_RESPONSE. Further
investigation is needed to determine whether this behaviour is exhibited by
all servers or just the one I happen to be testing against.