Handle discarding messages for multi-part responses
Basically this just moves the multi-part handling into the _ResponseBuffer.
This has a convenient side-effect for authentication in that the off
empty RESPONSE_VALUE sent before the AUTH_RESPONSE can be more easily
ignored as it's treated as part of a multi-part response by the buffer
which is cleared after successful authentication.