Move BaseServerQuerier up to the package level
It living in valve.source.a2s was a hold-over from before the
master server stuff existed. BSQ was never exposed as a public
API in its own right so this should be fine.

However, NoResponseError has been moved along with BSQ. This was
a public part of the interface. An alias exists for it in the A2S
module so not to break things. It should be removed in the future.